Reorganize repo for better imports
