+++ src/corpus_query_language/functions/functions.py
@@ -1,40 +1,9 @@
 import re
 import json
-import corpus_query_language.parser as parser
-import corpus_query_language.lexer as lexer
-import corpus_query_language.engine as engine
+import corpus_query_language.language.parser as parser
+import corpus_query_language.language.lexer as lexer
 
 
-class CQLEngine():
-	def findall(self, corpus:list[dict], query:str, verbose:bool=True,  debug:bool=False) -> list[tuple[int, int]]:
-		"""
-			This function checks if a query matches some text, and returns the start and end span.
-			:param query: a CQL query
-			:param corpus: the annotated text as a list of dictionnaries containing the annotations (lemma, pos, morph, word)
-			:return: a list of tuples with the start and end position.
-			"""
-		query_ast = build_grammar(debug=debug, query=query)
-		result = engine.parse_corpus(query_ast, corpus, mode="find", debug=debug)
-		if verbose:
-			print(f"\n---\nResults for query {query}:")
-			print(f"Ast: {query_ast}")
-			print(f"Spans: {result}")
-		return result
-
-
-	def match(self, corpus:list[dict], query:str, verbose:bool=True, debug:bool=False) -> bool:
-		"""
-		This function checks whether a query matches some text, and returns True or False
-		:param query: a CQL query
-		:param corpus: the annotated text as a list of dictionnaries containing the annotations (lemma, pos, morph, word)
-		:return: a boolean
-		"""
-		query_ast = build_grammar(debug=debug, query=query)
-		result = engine.parse_corpus(query_ast, corpus, mode="match", debug=debug)
-		if verbose:
-			print(f"\n---\nResults for query {query}:")
-			print(result)
-		return result
 
 
 def build_grammar(debug, query):
